feat: add auth, logger

+++ controllers/movies.js
@@ -1,6 +1,7 @@
 const Movie = require('../models/movie');
 const NotFoundError = require('../errors/not-found-err');
 const BadRequestError = require('../errors/bad-request-err');
+const ForbiddenError = require('../errors/forbidden-err');
 
 //возвращает все сохранённые текущим  пользователем фильмы
 module.exports.getMovies = (req, res, next) => {
@@ -11,8 +12,8 @@ module.exports.getMovies = (req, res, next) => {
 
 //создаёт фильм по переданным в теле параметрам
 module.exports.createMovie = (req, res, next) => {
-  const {country, director, duration, year, description, image, trailer, nameRU, nameEN, thumbnail, movieId } = req.body;
-  Movie.create({ country, director, duration, year, description, image, trailer, nameRU, nameEN, thumbnail, movieId, owner: req.user._id })
+  const {country, director, duration, year, description, image, trailerLink, nameRU, nameEN, thumbnail, movieId } = req.body;
+  Movie.create({ country, director, duration, year, description, image, trailerLink, nameRU, nameEN, thumbnail, movieId, owner: req.user._id })
     .then((movie) => res.send(movie))
     .catch((err) => {
       if (err.name === 'ValidationError') {
@@ -28,7 +29,9 @@ module.exports.createMovie = (req, res, next) => {
 module.exports.deleteMovieById = (req, res, next) => {
   Movie.findById(req.params._id)
     .then((movie) => {
+      console.log(!movie.owner.equals(req.user._id))
       if (!movie) throw new NotFoundError('Передан несуществующий id фильма');
+      if (!movie.owner.equals(req.user._id)) throw new ForbiddenError('Нельзя удалить чужой фильм');
       Movie.deleteOne(movie).then(() => res.send({ movie }));
     }).catch((err) => {
       if (err.name === 'CastError') {
